summary and checking out for the night

+++ get_audio_transcript.py
@@ -3,6 +3,18 @@ import speech_recognition as sr
 from tqdm import tqdm
 from multiprocessing.dummy import Pool
 
+def getExtension(path):
+        """
+        Gets the file extension from path
+
+        :param str path: Path of the file
+
+        :returns: File extension
+        :rtype: str
+        """
+        filename, file_extension = os.path.splitext(path)
+        return file_extension
+
 def get_audio_transcript(file_name, key_name):
     NUM_THREADS = 100 # Number of concurrent threads
     r = sr.Recognizer()
@@ -28,10 +40,15 @@ def get_audio_transcript(file_name, key_name):
     targDir = "parts"
     os.system("mkdir {}".format(targDir))
     os.system("ulimit -n 2048")
-    os.system('ffmpeg -i "{}" -acodec pcm_s16le -ac 1 -ar 8000 "{}".wav'.format(file_name, file_name))
-    command = 'ffmpeg -i "{}.wav" -f segment -segment_time 27 -c copy {}/out%09d.wav'.format(file_name, targDir)
-    os.system(command)
-    os.system('rm -rf "{}.wav"'.format(file_name))
+    ext = getExtension(file_name)
+    if ext == '.mp4':
+        os.system('ffmpeg -i "{}" -acodec pcm_s16le -ac 1 -ar 8000 "{}".wav'.format(file_name, file_name))
+        command = 'ffmpeg -i "{}.wav" -f segment -segment_time 27 -c copy {}/out%09d.wav'.format(file_name, targDir)
+        os.system(command)
+        os.system('rm -rf "{}.wav"'.format(file_name))
+    elif ext == '.wav':
+        command = 'ffmpeg -i "{}" -f segment -segment_time 27 -c copy {}/out%09d.wav'.format(file_name, targDir)
+        os.system(command)
     files = sorted(os.listdir('{}/'.format(targDir)))
     all_text = pool.map(transcribe, enumerate(files))
     pool.close()
